backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
jinja2==3.1.3
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic[email]==2.6.1
PyMuPDF==1.23.26
python-docx==1.1.0
sentence-transformers==2.5.1
faiss-cpu==1.7.4
numpy==1.26.4
requests==2.31.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
cryptography==42.0.2 